requirements.txt
supernotelib==v0.5.0
llama-index==0.11.4
PyYaml==6.0.1
google-generativeai>=0.7.2
ipykernel>=6.29.5
mashumaro==3.13.1
 syrupy==4.7.1
